speech/requirements.txt
openai
flask
requests-toolbelt
python-dotenv
cartesia
google-api-python-client
